+++ Assets/Scenes/NetworkMock/InGameMock.unity
@@ -1338,12 +1338,16 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 792960411302418462, guid: aea3ee96d9aafdc44987d473e0f5d27d, type: 3}
       propertyPath: NetworkedBehaviours.Array.size
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 792960411302418462, guid: aea3ee96d9aafdc44987d473e0f5d27d, type: 3}
       propertyPath: 'NetworkedBehaviours.Array.data[0]'
       value: 
       objectReference: {fileID: 542693326}
+    - target: {fileID: 792960411302418462, guid: aea3ee96d9aafdc44987d473e0f5d27d, type: 3}
+      propertyPath: 'NetworkedBehaviours.Array.data[1]'
+      value: 
+      objectReference: {fileID: 542693327}
     - target: {fileID: 5350015315830211674, guid: aea3ee96d9aafdc44987d473e0f5d27d, type: 3}
       propertyPath: m_SizeDelta.y
       value: 0
@@ -1472,6 +1476,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 248857c8872140cfb5d7ed6b59f56530, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &542693327 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 1939404891127187098, guid: aea3ee96d9aafdc44987d473e0f5d27d, type: 3}
+  m_PrefabInstance: {fileID: 542693323}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bb6bad39765d49ca98a5c273520c46f6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &644244725
@@ -8316,8 +8331,8 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1285741891474971311}
   serializedVersion: 2
-  m_LocalRotation: {x: 0.15260132, y: -0.8086604, z: 0.23953608, w: 0.5151735}
-  m_LocalPosition: {x: 20.002966, y: 0.77697897, z: 24.885712}
+  m_LocalRotation: {x: 0.16909811, y: 0.7703555, z: -0.22818965, w: 0.5708658}
+  m_LocalPosition: {x: -22.510273, y: 0.78999996, z: 24.146217}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
