+++ Assets/Scripts/InGame/Performances/PlayerEmotesPerformance.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 87ca69fec6fb41a1bebf84fc8d2018e5
+timeCreated: 1783679988
