+++ Assets/Scripts/InGame/Rules/GameStartLogic/IGameStartStrategy.cs
@@ -1,0 +1,7 @@
+namespace September.InGame.Rules
+{
+    public interface IGameStartStrategy
+    {
+        public void OnGameStarted();
+    }
+}
