+++ Assets/Scripts/InGame/Rules/PlayerKilled/IPlayerKilledUseCase.cs
@@ -1,0 +1,12 @@
+using Fusion;
+
+namespace September.InGame.Rules
+{
+    /// <summary>
+    /// プレイヤーをキルした際の処理を定義するユースケース
+    /// </summary>
+    public interface IPlayerKilledUseCase
+    {
+        public void ProcessKillEvent(PlayerRef killer, PlayerRef victim);
+    }
+}
