+++ Assets/Scripts/InGame/UI/Presenters/IGameStartedPresenter.cs
@@ -1,4 +1,4 @@
-namespace September.Common
+namespace September.InGame.UI.Presenters
 {
     public interface IGameStartedPresenter
     {
