+++ Assets/Scripts/InGame/UI/Presenters/IPlayerKilledPresenter.cs
@@ -1,6 +1,6 @@
 using Fusion;
 
-namespace September.Common
+namespace September.InGame.UI.Presenters
 {
     public interface IPlayerKilledPresenter
     {
